[#4] Extract seasons endpoint to dedicated resource file
